refactor: Extract CustomLink component
